Module Octez_node_config.Config_file

type chain_name = Tezos_base.TzPervasives.Distributed_db_version.Name.t

type blockchain_network = {

alias : string option;
(*
as given to --network, only for built-in networks
*)

genesis : Tezos_base.Genesis.t;

genesis_parameters : Tezos_base.Genesis.Parameters.t option;

chain_name : chain_name;

old_chain_name : chain_name option;

incompatible_chain_name : chain_name option;

sandboxed_chain_name : chain_name;

user_activated_upgrades : Tezos_base.User_activated.upgrades;

user_activated_protocol_overrides : Tezos_base.User_activated.protocol_overrides;

default_bootstrap_peers : string list;

dal_config : Tezos_crypto_dal.Cryptobox.Config.t;

}

val builtin_blockchain_networks : (string * blockchain_network) list

List of built-in networks with their alias.

val blockchain_network_encoding : blockchain_network Tezos_base.TzPervasives.Data_encoding.t

Data encoding for custom blockchain configuration.

type t = {

data_dir : string;

disable_config_validation : bool;

p2p : p2p;

rpc : rpc;

log : Tezos_base_unix.Logs_simple_config.cfg;

internal_events : Tezos_base.Internal_event_config.t option;

shell : Tezos_shell_services.Shell_limits.limits;

blockchain_network : blockchain_network;

metrics_addr : string list;

}

and p2p = {

expected_pow : float;

bootstrap_peers : string list option;

listen_addr : string option;

advertised_net_port : int option;

discovery_addr : string option;

private_mode : bool;

limits : Tezos_p2p_services.P2p_limits.t;

disable_mempool : bool;

enable_testchain : bool;

reconnection_config : Tezos_p2p_services.Point_reconnection_config.t;

disable_peer_discovery : bool;

}

and rpc = {

listen_addrs : string list;

external_listen_addrs : string list;

cors_origins : string list;

cors_headers : string list;

tls : tls option;

acl : Tezos_rpc_http_server.RPC_server.Acl.policy;

media_type : Tezos_rpc_http.Media_type.Command_line.t;

max_active_rpc_connections : Tezos_rpc_http_server.RPC_server.Max_active_rpc_connections.t;

enable_http_cache_headers : bool;

}

and tls = {

cert : string;

key : string;

}

val data_dir_env_name : string

val default_data_dir : string

val default_p2p_port : int

val default_rpc_port : int

val default_max_active_rpc_connections : Tezos_rpc_http_server.RPC_server.Max_active_rpc_connections.t

val default_p2p : p2p

val default_config : t

val update : ?disable_config_validation:bool -> ?data_dir:string -> ?min_connections:int -> ?expected_connections:int -> ?max_connections:int -> ?max_download_speed:int -> ?max_upload_speed:int -> ?binary_chunks_size:int -> ?peer_table_size:int -> ?expected_pow:float -> ?bootstrap_peers:string list option -> ?listen_addr:string -> ?advertised_net_port:int -> ?discovery_addr:string -> ?rpc_listen_addrs:string list -> ?external_rpc_listen_addrs:string list -> ?allow_all_rpc:Tezos_base.P2p_point.Id.addr_port_id list -> ?media_type:Tezos_rpc_http.Media_type.Command_line.t -> ?max_active_rpc_connections: Tezos_rpc_http_server.RPC_server.Max_active_rpc_connections.t -> ?metrics_addr:string list -> ?operation_metadata_size_limit: Tezos_shell_services.Shell_limits.operation_metadata_size_limit -> ?private_mode:bool -> ?disable_p2p_maintenance:bool -> ?disable_p2p_swap:bool -> ?disable_mempool:bool -> ?enable_testchain:bool -> ?cors_origins:string list -> ?cors_headers:string list -> ?rpc_tls:tls -> ?log_output:Tezos_base_unix.Logs_simple_config.Output.t -> ?log_coloring:bool -> ?synchronisation_threshold:int -> ?history_mode:Tezos_shell_services.History_mode.t -> ?network:blockchain_network -> ?latency:int -> ?enable_http_cache_headers:bool -> ?disable_context_pruning:bool -> ?storage_maintenance_delay:Tezos_shell_services.Storage_maintenance.delay -> t -> t Tezos_base.TzPervasives.tzresult Lwt.t

Update the given configuration. This is mostly used to apply the CLI arguments to the default config when starting or configuring the node. Arguments to this function roughly reflect the options in Shared_arg.t.

val to_string : t -> string

val read : string -> t Tezos_base.TzPervasives.tzresult Lwt.t

val write : string -> t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Write configuration file.

Also check whether the directory of the given filename is a data directory.

val resolve_listening_addrs : string -> Tezos_base.P2p_point.Id.t list Tezos_base.TzPervasives.tzresult Lwt.t

resolve_listening_addrs listening_addr parses listening_addr and returns a list of points. The default port is default_p2p_port. Fails if the address could not be parsed.

val resolve_discovery_addrs : string -> (Ipaddr.V4.t * int) list Tezos_base.TzPervasives.tzresult Lwt.t

resolve_discovery_addrs disco_addrs parses disco_addr and returns a list of points. disco_addrs must resolves to Ipv4 addresses otherwise it fails. The default port is default_discovery_port. Fails if the address could not be parsed.

val resolve_rpc_listening_addrs : string -> Tezos_base.P2p_point.Id.t list Tezos_base.TzPervasives.tzresult Lwt.t

resolve_rpc_listening_addrs rpc_addrs parses rpc_addr and returns a list of points. The default port is default_rpc_port. Fails if the address could not be parsed.

val resolve_metrics_addrs : ?default_metrics_port:int -> string -> Tezos_base.P2p_point.Id.t list Tezos_base.TzPervasives.tzresult Lwt.t

resolve_metrics_addrs ?default_metrics_port metrics_addr parses metrics_addr and returns a list of points). The default host is "localhost" and the default port is default_metrics_port. Fails if the address could not be parsed.

val resolve_bootstrap_addrs : string list -> (Tezos_base.P2p_point.Id.t * Tezos_base.P2p_peer.Id.t option) list Tezos_base.TzPervasives.tzresult Lwt.t

resolve_boostrap_addrs bs_addrs parses bs_addrs and returns for each addr a list of points. The default port is default_p2p_port. Fails if the address could not be parsed.

val rpc_encoding : rpc Tezos_base.TzPervasives.Data_encoding.t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val bootstrap_peers : t -> string list

Return p2p.bootstrap_peers if not None, network.default_bootstrap_peers otherwise.