Module Octez_node_config.Data_version

type Tezos_base.TzPervasives.error +=

| Invalid_data_dir of {

data_dir : string;

msg : string option;

}

Errors related to checks related to the data dir.

type Tezos_base.TzPervasives.error +=

| Could_not_read_data_dir_version of string

Default file names to store the information about the node's network identity, peers and configuration.

val default_identity_file_name : string

val default_config_file_name : string

val default_peers_file_name : string

type ensure_mode =

| Exists

| Is_bare

| Is_compatible

Defines the properties that must be satisfied by the call to ensure_data_dir.

Depending of the given mode, it will:

Exists: ensures that the directory exists, if not, creates a fresh directory by creating it and writing a version file.

Is_bare: ensures that, if the directory exists, it contains no other file than:

the version file;

the identity file;

the configuration file;

the peer list file; Otherwise, creates an fresh directory.

Is_compatible: similar to Exists but, additionally, will check the version compatibility.

val ensure_data_dir : ?mode:ensure_mode -> Tezos_base.Genesis.t -> string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

ensure_data_dir ~mode genesis data_dir ensures that a directory is valid with regards to the given mode. See ensure_mode for the mode's properties. By default, the mode is set to Is_compatible.

val upgrade_data_dir : data_dir:string -> Tezos_base.Genesis.t -> chain_name:Tezos_base.TzPervasives.Distributed_db_version.Name.t -> sandbox_parameters: (string * Tezos_base.TzPervasives.Data_encoding.json) option -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Upgrade data directory from an older version.

upgrade_data_dir dir genesis ~chain_name ~sandbox_parameters checks if an upgrade of the given data directory dir is available. If the data directory is upgradable then the upgrade is performed. Otherwise, nothing is done.

val store_dir : string -> string

store_dir dir is a directory within dir that the node uses for its store. In order for store_dir dir to be valid, dir must be a valid directory name.

val protocol_dir : string -> string

protocol_dir dir is a directory within dir that the node uses for its protocol. In order for protocol_dir dir to be valid, dir must be a valid directory name.

val lock_file : string -> string

lock_file dir is a file within dir that the node uses for its lock. In order for lock_file dir to be valid, dir must be a valid directory name.

val upgrade_status : data_dir:string -> bool Tezos_base.TzPervasives.tzresult Lwt.t

upgrade_status ~data_dir returns whether or not an upgrade is available. In addition to that, it prints a user friendly log.