Module Octez_node_config.Shared_arg

type net_config

type t = {

disable_config_validation : bool;

data_dir : string option;
(*
a directory where the node's configuration and state should be stored
*)

config_file : string;
(*
the config.json file to use
*)

network : net_config option;
(*
the P2P network to connect to
*)

connections : int option;

max_download_speed : int option;

max_upload_speed : int option;

binary_chunks_size : int option;

peer_table_size : int option;

expected_pow : float option;

peers : string list;

no_bootstrap_peers : bool;

listen_addr : string option;
(*
address to listen to connections from peers
*)

advertised_net_port : int option;
(*
port advertised for other peers to connect to
*)

discovery_addr : string option;

rpc_listen_addrs : string list;
(*
a list of addresses to listen to RPC requests on
*)

external_rpc_listen_addrs : string list;
(*
a list of addresses to listen to RPC requests on
*)

private_mode : bool;
(*
enables the private mode, see https://tezos.gitlab.io/user/node-configuration.html#private-node
*)

disable_p2p_maintenance : bool;
(*
If disable_p2p_maintenance is true the p2p layer will not open or close connections by itself. This flag is intended to be used for testing and debugging.
*)

disable_p2p_swap : bool;
(*
If disable_p2p_swap is true the p2p layer will neither initiate a swap of connections with its neighbors nor answer to a swap request. This flag is intended to be used for testing and debugging.
*)

disable_mempool : bool;

enable_testchain : bool;

cors_origins : string list;

cors_headers : string list;

rpc_tls : Config_file.tls option;

log_output : Tezos_base_unix.Logs_simple_config.Output.t option;

log_coloring : bool option;

bootstrap_threshold : int option;

history_mode : Tezos_shell_services.History_mode.t option;

synchronisation_threshold : int option;

latency : int option;

allow_all_rpc : Tezos_base.P2p_point.Id.addr_port_id list;
(*
a list of RPC listening addresses for which a full access should be granted
*)

media_type : Tezos_rpc_http.Media_type.Command_line.t;

max_active_rpc_connections : Tezos_rpc_http_server.RPC_server.Max_active_rpc_connections.t;

metrics_addr : string list;

operation_metadata_size_limit : Tezos_shell_services.Shell_limits.operation_metadata_size_limit option;
(*
maximum operation metadata size allowed to be stored on disk
*)

enable_http_cache_headers : bool option;
(*
Adds Cache-control header directives to RPC responses for queries that are relative to the head block.
*)

disable_context_pruning : bool option;

storage_maintenance_delay : Tezos_shell_services.Storage_maintenance.delay option;

}

This record represents a collection of command line arguments given when starting the node.

val process_command : unit Tezos_base.TzPervasives.tzresult Lwt.t -> unit Cmdliner.Term.ret

module Term : sig ... end

val read_config_file : t -> Config_file.t Tezos_base.TzPervasives.tzresult Lwt.t

val resolve_data_dir_and_config_file : ?data_dir:string -> ?config_file:string -> unit -> (string * Config_file.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val patch_network : ?cfg:Config_file.t -> Config_file.blockchain_network -> (Config_file.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Modify a node's network configuration.

This returns cfg modified according to the network argument. Default value for cfg is Config_file.default_config.

val patch_config : ?may_override_network:bool -> ?emit: (unit Tezos_base.TzPervasives.Internal_event.Simple.t -> unit -> unit Lwt.t) -> ?ignore_bootstrap_peers:bool -> ?cfg:Config_file.t -> t -> (Config_file.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Modify a node configuration.

This returns cfg modified according to command-line arguments. Default value for cfg is Config_file.default_config.

may_override_network specifies whether the network parameter is overridden, if already configured. Default value for may_override_network is false.

ignore_bootstrap_peers specifies whether the bootstrap peers, from cfg, are ignored or not. Default value for ignore_bootstrap_peers is false.

val read_and_patch_config_file : ?may_override_network:bool -> ?emit: (unit Tezos_base.TzPervasives.Internal_event.Simple.t -> unit -> unit Lwt.t) -> ?ignore_bootstrap_peers:bool -> t -> Config_file.t Tezos_base.TzPervasives.tzresult Lwt.t

read_and_patch_config_file ~may_override_network ~ignore_bootstrap_peers args Read the current config file and modify it accordingly to args.

may_override_network specifies whether the network parameter is overridden, if already configured. ignore_bootstrap_peers specifies whether the bootstrap peers, from the current config file, are ignored or not.

module Manpage : sig ... end